correctly attribute event type to removeLiquidity event
