+++ libs/common/src/services/xexchange/entities/index.ts
@@ -1,5 +1,6 @@
 export * from './pair';
 export * from './pair.metadata';
+export * from './xexchange.liquidity.event';
 export * from './xexchange.add.liquidity.event';
 export * from './xexchange.remove.liquidity.event';
 export * from './xexchange.swap.event';
