+++ libs/common/src/services/xexchange/entities/xexchange.remove.liquidity.event.ts
@@ -1,3 +1,9 @@
-import { XExchangeAddLiquidityEvent } from "./xexchange.add.liquidity.event";
+import { ElasticEvent, ElasticLog } from "../../indexer";
+import { XExchangePair } from "./pair";
+import { XExchangeLiquidityEvent } from "./xexchange.liquidity.event";
 
-export class XExchangeRemoveLiquidityEvent extends XExchangeAddLiquidityEvent { }
+export class XExchangeRemoveLiquidityEvent extends XExchangeLiquidityEvent {
+  constructor(event: ElasticEvent, log: ElasticLog, pair: XExchangePair) {
+    super(event, log, pair, "removeLiquidity");
+  }
+}
